[FINAL COMMIT] Continuous Card Flipping Logic Integrated
### Scripts

-CardManager.cs [Updated]
Before this change, when the user was clicking 3 different cards one-by-one in less than 1 second, then the last card's front side was remaining on the layout, which that was a bug. In order to fix that, a new list required to pick selected cards and queue them and handling them asynchronously. So now the player can click as fast as they can, and the cards will continue to appear and check the logic as required on the case document.